requirements.txt
Django
requests
django_ratelimit
python-telegram-bot
user-agents
gunicorn
whitenoise
